add saving data to automation dashboard (#5476)
