+++ .gitignore
@@ -12,6 +12,9 @@ mattermost-webapp.iml
 junit.xml
 coverage
 
+# Environment variables
+.env*
+
 # disable folders generated by Cypress
 e2e/node_modules
 e2e/cypress/screenshots
