+++ e2e/package.json
@@ -11,6 +11,7 @@
     "cypress-multi-reporters": "1.2.4",
     "cypress-plugin-tab": "1.0.5",
     "cypress-wait-until": "1.6.1",
+    "dotenv": "8.2.0",
     "express": "4.17.1",
     "lodash": "4.17.15",
     "lodash.intersection": "4.4.0",
@@ -25,9 +26,9 @@
     "mochawesome-report-generator": "5.1.0",
     "path": "0.12.7",
     "recursive-readdir": "2.2.2",
+    "reportportal-client": "5.5.0",
     "shelljs": "0.8.3",
     "start-server-and-test": "1.11.0",
-    "superagent": "5.2.2",
     "yargs": "15.3.1"
   },
   "scripts": {
