+++ e2e/run_tests.js
@@ -25,7 +25,7 @@
  *   HEADLESS=[boolean]     : Headless by default (true) or false to run on headed mode.
  *   BRANCH=[branch]        : Branch identifier from CI
  *   BUILD_ID=[build_id]    : Build identifier from CI
- *   TYPE=[type]            : Test type, e.g. "Daily", "PR"
+ *   TYPE=[type]            : Test type, e.g. "DAILY", "PR", RELEASE
  *   WEBHOOK_URL=[url]      : Webhook URL where to send test report
  *
  * Example:
@@ -41,180 +41,26 @@
 
 const os = require('os');
 
-const axios = require('axios');
 const chalk = require('chalk');
 const cypress = require('cypress');
 const fse = require('fs-extra');
-const intersection = require('lodash.intersection');
-const without = require('lodash.without');
 const {merge} = require('mochawesome-merge');
 const generator = require('mochawesome-report-generator');
-const shell = require('shelljs');
-const superagent = require('superagent');
 const argv = require('yargs').argv;
 
+const {getTestFiles, getSkippedFiles} = require('./utils/file');
+const {
+    generateDiagnosticReport,
+    generateShortSummary,
+    generateTestReport,
+    getServerInfo,
+    sendReport,
+    writeJsonToFile
+} = require('./utils/report');
 const {saveArtifacts} = require('./utils/save_artifacts');
-const users = require('./cypress/fixtures/users.json');
+const {saveDashboard} = require('./utils/dashboard');
 
-const MAX_FAILED_TITLES = 5;
-const TEST_DIR = 'cypress/integration';
-
-function getAllTests(results) {
-    const tests = [];
-    results.forEach((result) => {
-        result.tests.forEach((test) => tests.push(test));
-
-        if (result.suites.length > 0) {
-            getAllTests(result.suites).forEach((test) => tests.push(test));
-        }
-    });
-
-    return tests;
-}
-
-function generateStatsFieldValue(stats, failedFullTitles) {
-    let statsFieldValue = `
-| Key | Value |
-|:---|:---|
-| Passing Rate | ${stats.passPercent.toFixed(2)}% |
-| Duration | ${(stats.duration / (60 * 1000)).toFixed(2)} mins |
-| Suites | ${stats.suites} |
-| Tests | ${stats.tests} |
-| :white_check_mark: Passed | ${stats.passes} |
-| :x: Failed | ${stats.failures} |
-| :fast_forward: Skipped | ${stats.skipped} |
-`;
-
-    // If present, add full title of failing tests.
-    // Only show per maximum number of failed titles with the last item as "more..." if failing tests are more than that.
-    let failedTests;
-    if (failedFullTitles && failedFullTitles.length > 0) {
-        const re = /[:'"\\]/gi;
-        const failed = failedFullTitles;
-        if (failed.length > MAX_FAILED_TITLES) {
-            failedTests = failed.slice(0, MAX_FAILED_TITLES - 1).map((f) => `- ${f.replace(re, '')}`).join('\n');
-            failedTests += '\n- more...';
-        } else {
-            failedTests = failed.map((f) => `- ${f.replace(re, '')}`).join('\n');
-        }
-    }
-
-    if (failedTests) {
-        statsFieldValue += '###### Failed Tests:\n' + failedTests;
-    }
-
-    return statsFieldValue;
-}
-
-function generateShortSummary(report) {
-    const {results, stats} = report;
-    const tests = getAllTests(results);
-
-    const failedFullTitles = tests.filter((t) => t.fail).map((t) => t.fullTitle);
-    const statsFieldValue = generateStatsFieldValue(stats, failedFullTitles);
-
-    return {
-        stats,
-        statsFieldValue,
-    };
-}
-
-function writeJsonToFile(jsonObject, filename, dir) {
-    fse.writeJson(`${dir}/${filename}`, jsonObject).
-        then(() => console.log('Successfully written:', filename)).
-        catch((err) => console.error(err));
-}
-
-const grepCommand = (word = '') => {
-    // -r, recursive search on subdirectories
-    // -I, ignore binary
-    // -l, only names of files to stdout/return
-    // -w, expression is searched for as a word
-    return `grep -rIlw '${word}' ${TEST_DIR}`;
-};
-
-const grepFiles = (command) => {
-    return shell.exec(command, {silent: true}).stdout.
-        split('\n').
-        filter((f) => f.includes('spec.js'));
-};
-
-function getTestFiles() {
-    const {invert, group, stage} = argv;
-
-    const allFiles = grepFiles(grepCommand());
-
-    const stageFiles = [];
-    if (stage) {
-        const sc = grepCommand(stage.split(',').join('\\|'));
-        stageFiles.push(...grepFiles(sc));
-    }
-    const groupFiles = [];
-    if (group) {
-        const gc = grepCommand(group.split(',').join('\\|'));
-        groupFiles.push(...grepFiles(gc));
-    }
-
-    if (invert) {
-        // Return no test file if no stage and group, but inverted
-        if (!stage && !group) {
-            return [];
-        }
-
-        // Return all excluding stage files
-        if (stage && !group) {
-            return without(allFiles, ...stageFiles);
-        }
-
-        // Return all excluding group files
-        if (!stage && group) {
-            return without(allFiles, ...groupFiles);
-        }
-
-        // Return all excluding group and stage files
-        return without(allFiles, ...intersection(stageFiles, groupFiles));
-    }
-
-    // Return all files if no stage and group flags
-    if (!stage && !group) {
-        return allFiles;
-    }
-
-    // Return stage files if no group flag
-    if (stage && !group) {
-        return stageFiles;
-    }
-
-    // Return group files if no stage flag
-    if (!stage && group) {
-        return groupFiles;
-    }
-
-    // Return files if both in stage and group
-    return intersection(stageFiles, groupFiles);
-}
-
-function getSkippedFiles(initialTestFiles, platform, browser, headless) {
-    const platformFiles = grepFiles(grepCommand(`@${platform}`));
-    const browserFiles = grepFiles(grepCommand(`@${browser}`));
-    const headlessFiles = grepFiles(grepCommand(`@${headless ? 'headless' : 'headed'}`));
-
-    const initialSkippedFiles = platformFiles.concat(browserFiles, headlessFiles);
-    const skippedFiles = intersection(initialTestFiles, initialSkippedFiles);
-    const finalTestFiles = without(initialTestFiles, ...skippedFiles);
-
-    // Log which files were skipped
-    if (skippedFiles.length) {
-        console.log(chalk.cyan(`\nSkipped test files due to ${platform}/${browser} (${headless ? 'headless' : 'headed'}):`));
-
-        skippedFiles.forEach((file, index) => {
-            console.log(chalk.cyan(`- [${index + 1}] ${file}`));
-        });
-        console.log('');
-    }
-
-    return {skippedFiles, finalTestFiles};
-}
+require('dotenv').config();
 
 async function runTests() {
     const {
@@ -222,9 +68,9 @@ async function runTests() {
         BROWSER,
         BUILD_ID,
         CYPRESS_baseUrl, // eslint-disable-line camelcase
+        DASHBOARD_ENABLE,
         DIAGNOSTIC_WEBHOOK_URL,
         HEADLESS,
-        TEST_DASHBOARD_URL,
         TYPE,
         WEBHOOK_URL,
     } = process.env;
@@ -306,188 +152,37 @@ async function runTests() {
 
     // Generate short summary, write to file and then send report via webhook
     const summary = generateShortSummary(jsonReport);
+    console.log(summary);
     writeJsonToFile(summary, 'summary.json', mochawesomeReportDir);
 
     const result = await saveArtifacts(`../${mochawesomeReportDir}`, bucketFolder);
     if (result && result.success) {
-        console.log('Successfully uploaded artifacts.');
+        console.log('Successfully uploaded artifacts to S3.');
+        console.log(`https://${process.env.AWS_S3_BUCKET}.s3.amazonaws.com/${bucketFolder}/mochawesome.html`);
     }
 
     // Send test report to "QA: UI Test Automation" channel via webhook
-    if (WEBHOOK_URL) {
+    if (TYPE && WEBHOOK_URL) {
         const data = generateTestReport(summary, result && result.success, bucketFolder);
-        await sendReport('summary report', WEBHOOK_URL, data);
+        await sendReport('summary report to Community channel', WEBHOOK_URL, data);
     }
 
     // Send diagnostic report via webhook
-    const baseUrl = CYPRESS_baseUrl || 'http://localhost:8065'; // eslint-disable-line camelcase
-    const serverInfo = await getServerInfo(baseUrl);
-    if (serverInfo.enableDiagnostics && DIAGNOSTIC_WEBHOOK_URL) {
+    // Send on "DAILY" type only
+    if (TYPE === 'DAILY' && DIAGNOSTIC_WEBHOOK_URL) {
+        const baseUrl = CYPRESS_baseUrl || 'http://localhost:8065'; // eslint-disable-line camelcase
+        const serverInfo = await getServerInfo(baseUrl);
         const data = generateDiagnosticReport(summary, serverInfo);
-        await sendReport('diagnostic report', DIAGNOSTIC_WEBHOOK_URL, data);
+        await sendReport('test info for diagnostic analysis', DIAGNOSTIC_WEBHOOK_URL, data);
     }
 
-    if (TEST_DASHBOARD_URL) {
-        const data = {
-            report: jsonReport,
-            branch: BRANCH,
-            build: BUILD_ID,
-        };
-        await sendReport('dashboard data', TEST_DASHBOARD_URL, data);
+    // Save data to automation dashboard
+    if (DASHBOARD_ENABLE === 'true') {
+        await saveDashboard(jsonReport, BRANCH);
     }
 
     // exit with the number of failed tests
     process.exit(failedTests);
-}
-
-const result = [
-    {status: 'Passed', priority: 'none', cutOff: 100, color: '#43A047'},
-    {status: 'Failed', priority: 'low', cutOff: 98, color: '#FFEB3B'},
-    {status: 'Failed', priority: 'medium', cutOff: 95, color: '#FF9800'},
-    {status: 'Failed', priority: 'high', cutOff: 0, color: '#F44336'},
-];
-
-function generateTestReport(summary, isUploadedToS3, bucketFolder) {
-    const {BRANCH, BROWSER, BUILD_ID} = process.env;
-    const {statsFieldValue, stats} = summary;
-
-    let testResult;
-    for (let i = 0; i < result.length; i++) {
-        if (stats.passPercent >= result[i].cutOff) {
-            testResult = result[i];
-            break;
-        }
-    }
-
-    let jenkinsFields = [];
-    if (BUILD_ID) {
-        jenkinsFields = [
-            {
-                short: false,
-                title: 'Test Report',
-                value: `[Link to the report](https://build-push.internal.mattermost.com/job/mattermost-ui-testing/job/mattermost-cypress/${BUILD_ID}/artifact/mattermost-webapp/e2e/results/mochawesome-report/mochawesome.html)`,
-            }, {
-                short: false,
-                title: 'Screenshots',
-                value: `[Link to the screenshots](https://build-push.internal.mattermost.com/job/mattermost-ui-testing/job/mattermost-cypress/${BUILD_ID}/artifact/mattermost-webapp/e2e/results/mochawesome-report/screenshots/)`,
-            }, {
-                short: false,
-                title: 'New Commits',
-                value: `[Link to the new commits](https://build-push.internal.mattermost.com/job/mattermost-ui-testing/job/mattermost-cypress/${BUILD_ID}/changes)`,
-            }
-        ];
-    }
-
-    let awsS3Fields;
-    if (isUploadedToS3) {
-        awsS3Fields = [
-            {
-                short: false,
-                title: 'Test Report',
-                value: `[Link to the report](https://${process.env.AWS_S3_BUCKET}.s3.amazonaws.com/${bucketFolder}/mochawesome.html)`,
-            },
-        ];
-    }
-
-    return {
-        username: 'Cypress UI Test',
-        icon_url: 'https://www.mattermost.org/wp-content/uploads/2016/04/icon.png',
-        attachments: [{
-            color: testResult.color,
-            author_name: 'Cypress UI Test',
-            author_icon: 'https://www.mattermost.org/wp-content/uploads/2016/04/icon.png',
-            author_link: 'https://www.mattermost.com',
-            title: `Cypress UI Test Automation ${testResult.status}!`,
-            fields: [
-                {
-                    short: false,
-                    title: 'Environment',
-                    value: `Branch: **${BRANCH}**, Browser: **${BROWSER}**`,
-                },
-                ...jenkinsFields,
-                ...awsS3Fields,
-                {
-                    short: false,
-                    title: `Key metrics (required support: ${testResult.priority})`,
-                    value: statsFieldValue,
-                }
-            ],
-            image_url: 'https://pbs.twimg.com/profile_images/1044345247440896001/pXI1GDHW_bigger.jpg'
-        }],
-    };
-}
-
-function generateDiagnosticReport(summary, serverInfo) {
-    const {BRANCH, BUILD_ID} = process.env;
-
-    return {
-        username: 'Cypress UI Test',
-        icon_url: 'https://www.mattermost.org/wp-content/uploads/2016/04/icon.png',
-        attachments: [{
-            color: '#43A047',
-            author_name: 'Cypress UI Test',
-            author_icon: 'https://www.mattermost.org/wp-content/uploads/2016/04/icon.png',
-            author_link: 'https://community.mattermost.com/core/channels/ui-test-automation',
-            title: `Cypress UI Test Automation #${BUILD_ID}, **${BRANCH}** branch`,
-            fields: [{
-                short: false,
-                value: `Start: **${summary.stats.start}**\nEnd: **${summary.stats.end}**\nUser ID: **${serverInfo.sysadminId}**\nTeam ID: **${serverInfo.ad1TeamId}**`,
-            }],
-        }],
-    };
-}
-
-async function getServerInfo(baseUrl) {
-    const sysadmin = users.sysadmin;
-    const headers = {'X-Requested-With': 'XMLHttpRequest'};
-
-    const loginResponse = await axios({
-        method: 'post',
-        url: `${baseUrl}/api/v4/users/login`,
-        headers,
-        data: {login_id: sysadmin.username, password: sysadmin.password},
-    });
-
-    let cookieString = '';
-    const setCookie = loginResponse.headers['set-cookie'];
-    setCookie.forEach((cookie) => {
-        const nameAndValue = cookie.split(';')[0];
-        cookieString += nameAndValue + ';';
-    });
-
-    headers.Cookie = cookieString;
-
-    const configResponse = await axios({
-        method: 'get',
-        url: `${baseUrl}/api/v4/config`,
-        headers,
-    });
-
-    const teamResponse = await axios({
-        method: 'get',
-        url: `${baseUrl}/api/v4/teams/name/ad-1`,
-        headers,
-    });
-
-    return {
-        enableDiagnostics: configResponse.data.LogSettings.EnableDiagnostics,
-        sysadminId: loginResponse.data.id,
-        ad1TeamId: teamResponse.data.id,
-    };
-}
-
-async function sendReport(name, url, data) {
-    try {
-        const response = await superagent.post(url).send(data);
-
-        if (response.body) {
-            console.log(`Successfully sent ${name}.`);
-        }
-        return response;
-    } catch (er) {
-        console.log(`Something went wrong while sending ${name}.`, er);
-        return false;
-    }
 }
 
 runTests();
